Implemented sn not registered error

+++ lib/domain_model/patient_policy_model.dart
@@ -3,7 +3,7 @@ class PatientPolicyModel {
   int _numberOfRetries;
   String _pinType;
   bool _dataTimeLogging;
-  double _timeTillTheEnd;
+  int _timeTillTheEnd;
 
   String get errorHandler => _errorHandler;
 
@@ -13,12 +13,12 @@ class PatientPolicyModel {
 
   bool get dataTimeLogging => _dataTimeLogging;
 
-  double get timeTillTheEnd => _timeTillTheEnd;
+  int get timeTillTheEnd => _timeTillTheEnd;
 
   PatientPolicyModel.fromJson(Map<String, dynamic> json)
       : _errorHandler = json['errorHandler'],
         _numberOfRetries = json['numberOfRetries'],
         _pinType = json['pinType'],
         _dataTimeLogging = json['dataTimeLogging'],
-        _timeTillTheEnd = json['timeTillTheEnd'];
+        _timeTillTheEnd = json['timeTillTheEnd'] ;
 }
